git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6@8426 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ magick/pixel.c
@@ -3987,7 +3987,7 @@ MagickExport MagickBooleanType InterpolateMagickPixelPacket(const Image *image,
       pixel->opacity=0.0;
       pixel->index=0.0;
       count*=count;            /* number of pixels - square of size */
-      for (i=0; i < count; i++)
+      for (i=0; i < (ssize_t) count; i++)
       {
         AlphaBlendMagickPixelPacket(image,p+i,indexes+i,pixels,alpha);
         gamma=MagickEpsilonReciprocal(alpha[0]);
